style: update design for passphrase modals
